moved tests because mocking is all mocked up in old test file
